% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CampNormTbl9_1}
\alias{CampNormTbl9_1}
\title{A list of soil properties by soil type used in NicheMapR}
\format{
a data.frame with the following columns:
\describe{
 \item{Texture}{soil type (see [soilparams()])}
 \item{Silt}{proportion of silt (0-1)}
 \item{Clay}{proportion of clay (0-1)}
 \item{airentry}{}
 \item{b}{}
 \item{Ks}{}
 \item{field}{}
 \item{wilting}{}
}
}
\usage{
CampNormTbl9_1
}
\description{
A list of soil properties by soil type used in NicheMapR
}
\keyword{datasets}
