% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microclimatemodel.R
\name{canopyrad}
\alias{canopyrad}
\title{Two-stream radiation}
\usage{
canopyrad(microin, i)
}
\arguments{
\item{microin}{an object of class `microin` as returned by [modelin()]}

\item{i}{time step for wich radiation is required}
}
\value{
a list of downward and upward diffuse and direct radiation streams.
}
\description{
Calculates upward and downward shortwave direct and diffuse radiation
}
\details{
Applies the variant of the Dickenson-Sellers two-stream model
described by Yuan et al (2017). A demonstration of its application is given
in the accompanying vignette.
}
