% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microclimatemodel.R
\name{canwind}
\alias{canwind}
\title{Wind profile coefficient}
\usage{
canwind(microin)
}
\arguments{
\item{microin}{an object of class `microin` as returned by [modelin()]}
}
\value{
a vector of wind speed coefficients (0-1)
}
\description{
Creates a vector by which wind speed at the top of the canopy
is multiplied to derive the vertical wind-height profile.
}
\details{
See accompanying vignette for application of this function.
}
