% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treetrunkmodel.R
\name{crossectionplot}
\alias{crossectionplot}
\title{plots a cross-section of tree trunk temperatures}
\usage{
crossectionplot(modin, modout, i, zlim = c(NA, NA))
}
\arguments{
\item{modin}{an object of class `treein` as returned by [modelprep()].}

\item{modout}{an object of class `treemodout` as returned by [runtreetrunk()].}

\item{i}{the time-increment for which to plot a cross-section of temperatures.}
}
\description{
for a specified time period, plots a raster representing the
cross-sectional temperatures of the tree-trunk.
time-series of temperatures of the outer layer of the tree-trunk
}
\details{
Requires the packages `fields` and `raster` to be installed as
temperatures within each segement and layer, as returned by [runtreetrunk()]
are interpolated to a 100 x 100 pixel resolution using the `fields` prior to
plotting. See accompanying vignette for a demonstration of this function..
}
