% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_microclimate.R
\name{.calcH}
\alias{.calcH}
\title{Internal function for calculating sensible heat below canopy}
\usage{
.calcH(microin, miconeout, ws, i)
}
\description{
Internal function for calculating sensible heat below canopy
}
