% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_microclimate.R
\name{.CalcTh}
\alias{.CalcTh}
\title{Internal function for calculating temperature and humidity and top of canopy}
\usage{
.CalcTh(microin, HT, i)
}
\description{
Internal function for calculating temperature and humidity and top of canopy
}
