% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_microclimate.R
\name{.cank}
\alias{.cank}
\title{Internal function for calculating canopy extinction coefficient}
\usage{
.cank(z, x)
}
\description{
Internal function for calculating canopy extinction coefficient
}
