% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_microclimate.R
\name{.canopyradp}
\alias{.canopyradp}
\title{Internal function for calculate parameters of two stream canopy radiation model}
\usage{
.canopyradp(vegp, weather, gref)
}
\description{
Internal function for calculate parameters of two stream canopy radiation model
}
