% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_microclimate.R
\name{.clearskyrad}
\alias{.clearskyrad}
\title{Internal function for calculating clear sky radiation}
\usage{
.clearskyrad(zenith)
}
\description{
Internal function for calculating clear sky radiation
}
