% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_microclimate.R
\name{.Langrangian}
\alias{.Langrangian}
\title{Internal function for running L-Nf-T Langrangian model}
\usage{
.Langrangian(HT, microin, miconeout, h)
}
\description{
Internal function for running L-Nf-T Langrangian model
}
