% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_microclimate.R
\name{.LangrangianR}
\alias{.LangrangianR}
\title{#' Internal function for iterating L-Nf-T Langrangian model}
\usage{
.LangrangianR(HT, microin, miconeout, h, R = 5)
}
\description{
#' Internal function for iterating L-Nf-T Langrangian model
}
