% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_treetrunk.R
\name{.layerexchange}
\alias{.layerexchange}
\title{Internal function for calculating layer heat exchange}
\usage{
.layerexchange(modin, i)
}
\description{
Internal function for calculating layer heat exchange
}
