% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_treetrunk.R
\name{.limtm}
\alias{.limtm}
\title{Internal function for setting limits to tree trunk temperatures}
\usage{
.limtm(tm, tmin, tmax)
}
\description{
Internal function for setting limits to tree trunk temperatures
}
