% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_microclimate.R
\name{.NicheMapRin}
\alias{.NicheMapRin}
\title{Internal function for creating NIcheMapR input}
\usage{
.NicheMapRin(microin, soilparams, twostreamp)
}
\description{
Internal function for creating NIcheMapR input
}
