% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_treetrunk.R
\name{.penman}
\alias{.penman}
\title{Internal function = temperatures derived using Penman-Monteith equation}
\usage{
.penman(modin, tsu, fwet)
}
\description{
Internal function = temperatures derived using Penman-Monteith equation
}
