% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_treetrunk.R
\name{.segmentexchange}
\alias{.segmentexchange}
\title{Internal function for calculating segment heat exchange}
\usage{
.segmentexchange(modin, i)
}
\description{
Internal function for calculating segment heat exchange
}
