% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_treetrunk.R
\name{.setlims}
\alias{.setlims}
\title{internal function for setting limits when splining weather data}
\usage{
.setlims(xs, xl)
}
\description{
internal function for setting limits when splining weather data
}
