% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_microclimate.R
\name{.ShadeF}
\alias{.ShadeF}
\title{Internal function for calculating effective shade factor}
\usage{
.ShadeF(microin, twostreamp)
}
\description{
Internal function for calculating effective shade factor
}
