% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_treetrunk.R
\name{.smoothsegments}
\alias{.smoothsegments}
\title{internal function for smoothing segment heats to avoid temperature yo-yo}
\usage{
.smoothsegments(modin, i)
}
\description{
internal function for smoothing segment heats to avoid temperature yo-yo
}
