% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_microclimate.R
\name{.solarcoef}
\alias{.solarcoef}
\title{Internal function for calculating solar coefficient}
\usage{
.solarcoef(solar, groundp)
}
\description{
Internal function for calculating solar coefficient
}
