% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_microclimate.R
\name{.stomcond}
\alias{.stomcond}
\title{Internal function for calculating stomatal conductance}
\usage{
.stomcond(Rsw, gsmax, q50 = 100)
}
\description{
Internal function for calculating stomatal conductance
}
