% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_treetrunk.R
\name{.treeouter}
\alias{.treeouter}
\title{Internal function for calculating temperature of outer layer}
\usage{
.treeouter(modin, EB, i)
}
\description{
Internal function for calculating temperature of outer layer
}
