% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_treetrunk.R
\name{.trunkradabs}
\alias{.trunkradabs}
\title{Internal function for calculating radiation absorbed}
\usage{
.trunkradabs(modin, i)
}
\description{
Internal function for calculating radiation absorbed
}
