% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{groundp}
\alias{groundp}
\title{#'
A list of soil and ground parameters required for running the model}
\format{
a list with the following elements:
\describe{
 \item{slope}{slope of gorund surface (decimal degrees)}
 \item{aspect}{aspect of ground surface (decimal degrees from N)}
 \item{gref}{ground refklectance (0-1)}
 \item{soil type}{soil type (see [soilparams()])}
}
}
\usage{
groundp
}
\description{
#'
A list of soil and ground parameters required for running the model
}
\keyword{datasets}
