% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microclimatemodel.R
\name{groundtemp}
\alias{groundtemp}
\title{Calculates ground surface temperature}
\usage{
groundtemp(microin, soilparams)
}
\arguments{
\item{microin}{an object of class `microin` as returned by [modelin()]}

\item{soilparams}{a data.frame of soil parameters (see inbuilt dataset `soilparams`)}
}
\value{
Adds soil surface temperatures to [modelin()]
}
\description{
Applies NicheMapR model to derive ground surface temperatures
}
\details{
See accompanying vignette for application of this function. Requires
NicheMapR (github.com/mrke/NicheMapR)
}
