% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4definitions.R
\docType{class}
\name{microin}
\alias{microin}
\title{microin S4 class
eg used in modelin function}
\description{
microin S4 class
eg used in modelin function
}
\section{Slots}{

\describe{
\item{\code{weather}}{data.frame.}

\item{\code{prec}}{numeric.}

\item{\code{vegp}}{list.}

\item{\code{groundp}}{list.}

\item{\code{loc}}{numeric.}

\item{\code{twostreamp}}{list.}

\item{\code{reqhgt}}{numeric.}
}}

