% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{microout}
\alias{microout}
\title{Microclimate model outputs at 5 m above
ground at Caerthillian Cove, Lizard, Cornwall (49.96807N, 5.215668W) for the
period 16th-30th June 2017, assuming a 10 m tall canopy and with the following
slots.}
\format{
an object of class microout with the following slots:
\describe{
  \item{Microclimate}{A list of air and foliage temperatures and relative humidities}
  \item{Radiation}{A list of upward and downward radiative fluxes}
  \item{InputWeather}{A data.frame of weather variables provided as inputs}
}
}
\usage{
microout
}
\description{
Microclimate model outputs at 5 m above
ground at Caerthillian Cove, Lizard, Cornwall (49.96807N, 5.215668W) for the
period 16th-30th June 2017, assuming a 10 m tall canopy and with the following
slots.
}
\keyword{datasets}
