% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treetrunkmodel.R
\name{microspline}
\alias{microspline}
\title{Spline interpolates microclimate input}
\usage{
microspline(modin, secs = 60)
}
\arguments{
\item{modin}{an object of class `treein` as returned by [modelprep()].}

\item{secs}{time-interval (seconds) of required output data}
}
\value{
an S4 object of class `treein` with higher temperal resolution data
}
\description{
Spline interpolates microclimate input to tree trunk model to shorter time intervals.
}
\details{
See accompanying vignette for application of this function.
}
