% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microclimatemodel.R
\name{modelin}
\alias{modelin}
\title{Creates microclimate model input}
\usage{
modelin(weather, prec, vegp, groundp, lat, long, reqhgt, n = 100)
}
\arguments{
\item{weather}{a data,frame of hourly weather (see details).}

\item{prec}{a vector of daily precipitation (see details).}

\item{vegp}{a list of vegetation parameters (see details).}

\item{groundp}{a list of soil parameters (see details).}

\item{lat}{latitude (decimal degrees)}

\item{long}{longitude (decimal degrees, -ve west of Greenwich meridian)}

\item{reqhgt}{height above ground for which model should be run (m)}

\item{n}{number of vertical layers over which to run model (default 100)}
}
\value{
an S4 object of class `microin` used as an input to the microclimate model
}
\description{
Creates an S4 object of class `microin` used as an input to
the microclimate model
}
\details{
See accompanying vignette for application of this function. The inbuilt
datasets `weather`, `prec`, `vegp` and `groundp` give examples of the components
used for building the model input. User-created inputs should follow the same format
and units as in these inbuilt datasets.
}
