% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treetrunkmodel.R
\name{modelprep}
\alias{modelprep}
\title{Creates tree trunk model input}
\usage{
modelprep(microout, treeparams)
}
\arguments{
\item{microout}{an object of class `microout` as returned by [runmicro()].}

\item{treeparams}{a list of tree parameters required for running the tree trunk model (see details).}
}
\value{
an S4 object of class `treein` used as an input to the tree trunk model
}
\description{
Creates an S4 object of class `treein` used as an input to
the tree trunk model
}
\details{
See accompanying vignette for application of this function. The inbuilt
dataset `treeparams` give an examples of the components used for building the
model input. User-created inputs should follow the same format and units as in these inbuilt datasets.
}
