% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microclimatemodel.R
\name{plotmicro}
\alias{plotmicro}
\title{Plot model results}
\usage{
plotmicro(microout, var = "Tair")
}
\arguments{
\item{microout}{an object of class `microout` as returned by [runmicro()]}

\item{var}{one of `Tair` (air temperature), `Tfoliage` (foliage temperature),
`Relhum` (relative humidity), `Windspeed` (wind speed) `Rdirdown` (Downward direct radiation), `Rdifdown`
(downward diffuse radiation), `Rdifup` (upward shortwave radiation, assumed
entirely diffuse), `Rlwdown` (downward longwave radiation), Rlwup (upward
longwave radiation). Default: 'Tair'}
}
\description{
This function is used plot time-series of model outputs
}
\details{
See accompanying vignette for application of this function.
}
