% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prec}
\alias{prec}
\title{A vector of daily rainfall at Caerthillian Cove, Lizard, Cornwall
(49.96807N, 5.215668W) for the period 16th-30th June 2017}
\format{
A vector of daily rainfall (mm/day)
}
\usage{
prec
}
\description{
A vector of daily rainfall at Caerthillian Cove, Lizard, Cornwall
(49.96807N, 5.215668W) for the period 16th-30th June 2017
}
\keyword{datasets}
