% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microclimatemodel.R
\name{runmicro}
\alias{runmicro}
\title{Runs microclimate model}
\usage{
runmicro(microin, interval = 300, iter = 5, tomax = NA)
}
\arguments{
\item{microin}{an object of class `microin` as returned by [modelin()]}

\item{interval}{time increment in seconds over which to run model}

\item{iter}{Number of iterations over which Langrangian microclimate model
is run witin each time-step to derive stable values. Default: 5}

\item{tomax}{time step to which to run model. If `NA` model is run for entire
duration}
}
\value{
an S4 object of class `microout` containing the folowing slots:

`Microclimate` a list containing vectors of air and foliage temperatures (deg C)
and relative humidifies (percentage) for each time increment.

`Radiation`  a list containing vectors of downard direct, downward diffuse,
upward diffuse and upward and downward longwave radiation (W/m^2)

`Inputweather` a data.frame of pline-interpolated input weather
variables used to run the model.
}
\description{
This function runs the full microclimate model.
}
\details{
See accompanying vignette for applictaion of the model and a detailed
description of model outputs.
}
