% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microclimatemodel.R
\name{runmicroone}
\alias{runmicroone}
\title{Runs one step of microclimate model}
\usage{
runmicroone(
  microin,
  miconeout,
  ws,
  i = 1,
  iter = 5,
  plotout = TRUE,
  plotvar = "tair"
)
}
\arguments{
\item{microin}{an object of class `microin` as returned by [modelin()].}

\item{miconeout}{a list of model outputs fromthe previous time-step or as
returned by [setinit()]}

\item{ws}{a vector of wind coeffients as returned by [canwind()].}

\item{i}{which time-increment of the model to run. Default 1 (the first time-step).}

\item{iter}{Number of iterations over which Langrangian microclimate model
is run witin each time-step to derive stable values. Default: 5}

\item{plotout}{optional logical indicating whether to plot outputs. Default: TRUE}

\item{plotvar}{One of `tair` (air temperature, the default), `tleaf` (leaf
temperature) or `rh` (relative humidity). Ignored if `plotout = FALSE`.}
}
\value{
a list containing vectors of air and leaf temperatures and
relative humidifies (see details)
}
\description{
This function runs one step of the microclimate model
}
\details{
Three vectors of length `n` are returned, the `ith` element of
which corresponds to a height (i / n) x canopy height above ground. An example
of the application of this function is shown in the accompanying vignette.
}
