% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treetrunkmodel.R
\name{runtreetrunk}
\alias{runtreetrunk}
\title{Runs tree trunk model}
\usage{
runtreetrunk(modin, fwet = 0.1, tomax = NA, daysinit = 1, smoothseg = FALSE)
}
\arguments{
\item{modin}{an object of class `treein` as returned by [modelprep()].}

\item{fwet}{fraction of tree trunk surface acting like a saturated water surface.}

\item{tomax}{time step to which to run model. If `NA` model is run for entire
duration}

\item{daysinit}{number of days to use as a burn-in when starting the model (see details).}

\item{smoothseg}{- boolean whether to run smoothing between segements}
}
\value{
an object of class `treemodout` - a list of the following: (i) an array of temperatures for each layer,
segment and time-interval; (ii) a POSIXlt object of the corresponding times
for each time-interval.
}
\description{
Runs a single step of the tree trunk model
}
\details{
this function calls [treeonestep()] repeatedly for each time-increment.
To ensure plausible starting values, temperatures in each segment and layer are
set to air temperature and the model run for `daysinit` days to introduce a
realistic temperature profile. The model is then run from the beginning. See accompanying
vignette for a demonstration of this function.
}
