% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microclimatemodel.R
\name{setinit}
\alias{setinit}
\title{Create model inputs for first time-step}
\usage{
setinit(microin)
}
\arguments{
\item{microin}{an object of class `microin` as returned by [modelin()]}
}
\value{
a list containing vectors of air and leaf temperatures and
relative humidifies (see details)
}
\description{
The microclimate model is run sequentially in time-increments, using
as an input, the output from the previous time-step. This function creates the
input for the first time-step
}
\details{
Three vectors of length `n` are returned, the `ith` element of
which corresponds to a height (i / n) x canopy height above ground. An example
of the application of this function is shown in the accompanying vignette.
}
