% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soilparams}
\alias{soilparams}
\title{A table of soil parameters for different soil types}
\format{
A data frame with the following columns:
\describe{
  \item{Soil.type}{description of soil type}
  \item{Smax}{Volumetric water content at saturation (m^3 / m^3)}
  \item{Smin}{Residual water content (m^3 / m^3)}
  \item{Ksat}{Saturated hydraulic conductivity (kg / m^3 / day)}
  \item{Vq}{Volumetric quartz content of soil}
  \item{Vm}{Volumetric mineral content of soil}
  \item{Vo}{Volumetric organic content of soil}
  \item{Mc}{Mass fraction of clay}
  \item{rho}{Soil bulk density (Mg / m^3)}
  \item{b}{Shape parameter for NicheMapR soil moisture model (dimensionless, > 1)}
}
}
\source{
: \url{https://onlinelibrary.wiley.com/doi/full/10.1002/ird.1751}
}
\usage{
soilparams
}
\description{
A table of soil parameters for different soil types
}
\keyword{datasets}
