% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microclimatemodel.R
\name{solarposition}
\alias{solarposition}
\title{Calculates solar position}
\usage{
solarposition(tme, lat, long)
}
\arguments{
\item{tme}{POSIXlt object of times in UTC}

\item{lat}{latitude (decimal degrees)}

\item{long}{(decimal degrees, -ve west of Greenwich meridian)}
}
\value{
a list of zenith and azimuth angles (decimal degrees)
}
\description{
Calculates the solar zenith and azimuth angles
}
\examples{
# At noon at Porthleven, Cornwall on 21st Jun
tme <- as.POSIXlt(0, origin = "2023-06-21 12:00", tz = "UTC")
solarposition(tme, 50.08, -5.31)
# Hourly at Porthleven, Cornwall on 21st Jun
tme <- as.POSIXlt(c(0:23) * 3600, origin = "2023-06-21 00:00", tz = "UTC")
sp <- solarposition(tme, 50.08, -5.31)
sa <- 90 - sp$zenith
plot(sa ~ as.POSIXct(tme), pch = 19, cex = 4, col = "yellow",
     xlab = "Time", ylab = "Solar altitude")
}
