% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treetrunkmodel.R
\name{timeseriesplot}
\alias{timeseriesplot}
\title{plots a time-series of tree-trunk temperatures}
\usage{
timeseriesplot(treemodout, fun = max, col = "black")
}
\arguments{
\item{treemodout}{an object of class `treemodout` as returned by [runtreetrunk()].}

\item{fun}{one of min, max or mean (see details).}

\item{col}{color of plot line}
}
\description{
plots a time-series of temperatures of the outer layer of the tree-trunk
}
\details{
If `fun = mean`, within each time-step, the mean temperature of the outer ring
is plotted. If `fun = min` or `fun = max`, within each time-step, the temperatures in
the coldest or warmest segment are selected for plotting. See accompanying vignette for
a demonstration of this function.
}
