% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4definitions.R
\docType{class}
\name{treein}
\alias{treein}
\title{treein S4 class
input class from microspline or tree trunk modelling}
\description{
treein S4 class
input class from microspline or tree trunk modelling
}
\section{Slots}{

\describe{
\item{\code{Microclimate}}{list.}

\item{\code{Radiation}}{list.}

\item{\code{InputWeather}}{data.frame.}

\item{\code{Treeparams}}{list.}

\item{\code{FixedParams}}{list.}

\item{\code{treetemps}}{matrix.}
}}

