% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treetrunkmodel.R
\name{treeonestep}
\alias{treeonestep}
\title{Runs one step of tree trunk model}
\usage{
treeonestep(modin, i, fwet = 0.1)
}
\arguments{
\item{modin}{an object of class `treein` as returned by [modelprep()].}

\item{i}{time-step of model to run.}

\item{fwet}{fraction of tree trunk surface acting like a saturated water surface.}
}
\value{
an S4 object of class `treein` used as an input to the next iteration of the tree trunk model
}
\description{
Runs a single step of the tree trunk model
}
\details{
See accompanying vignette for application of this function.
}
