% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{treeparams}
\alias{treeparams}
\title{A list of tree parameters required for running the tree trunk model}
\format{
a list with the following elements:
\describe{
 \item{nlayers}{total plant index value for the canopy}
 \item{nsegs}{ratio of vertical to horizontal projections of leaf foliage}
 \item{tref}{tree trunk reflectance (0-1)}
 \item{tem}{tree trunk emissivity (0-1}
 \item{radius}{tree trunk radius (m)}
 \item{K}{a vector of thermal conductivities of each layer (W/m/K)}
 \item{cph}{a vector of volumeric specific heat capacities of each layer (MJ/m^3)}
}
}
\usage{
treeparams
}
\description{
A list of tree parameters required for running the tree trunk model
}
\keyword{datasets}
