% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vegp}
\alias{vegp}
\title{A list of vegetation parameters required for running the microclimate model}
\format{
a list with the following elements:
\describe{
 \item{pait}{total plant index value for the canopy}
 \item{vegx}{ratio of vertical to horizontal projections of leaf foliage}
 \item{clump}{fraction of radiation passing through larger gaps in the canopy (0-1)}
 \item{lref}{leaf reflectance (0-1)}
 \item{ltra}{leaf transmittance (0-1)}
 \item{veghgt}{canopy height (m)}
 \item{em}{leaf emissivity (0-1)}
 \item{gsmax}{maximum stomatal conductance of leaves (mol / m^2 / s)}
 \item{leafd}{leaf diameter (m)}
 \item{paii}{individual leaf area index values for each canopy layer, here set to NA as skew and spread are used to estimate this when running the model}
 \item{skew}{degree of skew towards top of canopy in canopy foliage (0-10, see [PAIgeometry())}
 \item{spread}{degree of spread in canopy foliage (0-100, see [PAIgeometry())}
}
}
\usage{
vegp
}
\description{
A list of vegetation parameters required for running the microclimate model
}
\keyword{datasets}
