% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microclimatemodel.R
\name{weatherspline}
\alias{weatherspline}
\title{spline interpolate weather inputs}
\usage{
weatherspline(microin, secs = 300)
}
\arguments{
\item{microin}{an object of class `microin` as returned by [modelin()]}

\item{secs}{time interval in seconds of spline-interpolated data. Default: 300}
}
\value{
an object of class `microin` contained a data.frame of spline interpolated weather
variables and ground surface tmeperatures.
}
\description{
Spline interpolates hourly weather inputs to shorter time intervals.
}
\details{
See accompanying vignette for application of this function.
}
